Refactor user router
